🚧 resolve PR comments
